refactor(web): adopt react hook form and split book expand modal into sub-components
- Extract BookEntryView with view/edit toggle: read-only display by default,
  full RHF form on Edit button click; status remains immediate via pill
- Rewrite BookCreateForm to manage its own form state via RHF, removing
  dependency on hook-provided draft state
- Simplify useBookExpandModal to animation shell only: layout animation refs,
  handleClose, and handleCreate; all edit/deletion/status state moved to BookEntryView
- BookExpandModal becomes a thin orchestration shell (~85 lines vs ~540)
